Day315✅/requirements.txt
fastapi
uvicorn
python-dotenv
transformers
pydantic
